refactor: move footer to reusable component
